ai/requirements.txt
Flask==2.2.2
ipykernel==6.16.1
ipython==8.5.0
ipywidgets==8.0.1
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
pandas==1.4.3
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
scikit_learn==1.1.2
